Cylinder attached to camera
An interesting solution where the cylinder is attached to the camera. The issue is that past the wall there is still the pipe darkening overlay which makes it look like the hole is really big. The other problem is if I make the cylinder small in scale size, then it doesnt look as realistic as well. Two ways to do it:

1) either make it able that the darkening doesnt happen between the wall and the player
2) or make the pipe scale up/down depending on how thick the objects that block the view are